src/test/javascript/spec/app/entities/additional-user-info/additional-user-info.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { AdditionalUserInfoService } from 'app/entities/additional-user-info/additional-user-info.service';
import { IAdditionalUserInfo, AdditionalUserInfo } from 'app/shared/model/additional-user-info.model';

describe('Service Tests', () => {
  describe('AdditionalUserInfo Service', () => {
    let injector: TestBed;
    let service: AdditionalUserInfoService;
    let httpMock: HttpTestingController;
    let elemDefault: IAdditionalUserInfo;
    let expectedResult: IAdditionalUserInfo | IAdditionalUserInfo[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(AdditionalUserInfoService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new AdditionalUserInfo(0, 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a AdditionalUserInfo', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new AdditionalUserInfo()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a AdditionalUserInfo', () => {
        const returnedFromService = Object.assign(
          {
            phone: 'BBBBBB',
            profilePicture: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of AdditionalUserInfo', () => {
        const returnedFromService = Object.assign(
          {
            phone: 'BBBBBB',
            profilePicture: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a AdditionalUserInfo', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
